refactor: package 구조 적용
