+++ src/main/kotlin/carRace/domain/CarLapResult.kt
@@ -1,4 +1,4 @@
-package carRace
+package carRace.domain
 
 data class CarLapResult(
     val lap: Int,
